Module posish::io

I/O operations.

Structs

PollFd

pollfd

PollFdVec

A Vec of libc::pollfd.

Functions

dup

dup(fd)

fionread

ioctl(fd, FIONREAD).

is_read_write

Returns a pair of booleans indicating whether the file descriptor is readable and/or writeable, respectively. Unlike is_file_read_write, this correctly detects whether sockets have been shutdown, partially or completely.

isatty

isatty(fd)

socketpair_stream

socketpair(domain, SOCK_STREAM | SOCK_CLOEXEC, protocol)

ttyname

ttyname_r(fd)