Trait posish::path::Arg

pub trait Arg {
    fn as_str(&self) -> Result<&str>;

    fn to_string_lossy(&self) -> Cow<'_, str>;

    fn as_c_str(&self) -> Result<Cow<'_, CStr>>;

    fn as_maybe_utf8_bytes(&self) -> &[u8];
}

A trait for passing path arguments. This is similar to AsRef<Path>, but is implemented for more kinds of strings and can convert into more kinds of strings.

Required methods

fn as_str(&self) -> Result<&str>

Return a view of this string as a string slice.

fn to_string_lossy(&self) -> Cow<'_, str>

Return a potentially-lossy rendering of this string as a Cow<str>.

fn as_c_str(&self) -> Result<Cow<'_, CStr>>

Return a view of this string as a maybe-owned CStr.

fn as_maybe_utf8_bytes(&self) -> &[u8]

Return a view of this string as a byte slice.